Added slug validation to parade controller

+++ backend/src/api/parades/parades.controller.ts
@@ -5,11 +5,12 @@ import {
   Body,
   Patch,
   Param,
-  Delete,
+  Delete, BadRequestException,
 } from '@nestjs/common';
 import { ParadesService } from './parades.service';
 import { CreateParadeDto } from './dto/create-parade.dto';
 import { UpdateParadeDto } from './dto/update-parade.dto';
+import { Public } from '../../constants';
 
 @Controller('parades')
 export class ParadesController {
@@ -25,18 +26,22 @@ export class ParadesController {
     return this.paradesService.findAll();
   }
 
+  @Public()
   @Get(':id')
   findOne(@Param('id') id: string) {
+    if (!Number(id)) throw new BadRequestException();
     return this.paradesService.findOne(+id);
   }
 
   @Patch(':id')
   update(@Param('id') id: string, @Body() updateParadeDto: UpdateParadeDto) {
+    if (!Number(id)) throw new BadRequestException();
     return this.paradesService.update(+id, updateParadeDto);
   }
 
   @Delete(':id')
   remove(@Param('id') id: string) {
+    if (!Number(id)) throw new BadRequestException();
     return this.paradesService.remove(+id);
   }
